Issue #19 : Enhancement to upload an existing file
This PR add the possibility to configure the portlet behaviour when
uploading an existing file by adding a portlet preference property. It's
a first step before developping an UI to let the choice of the user when
the problem appear. This PR is more a first step but let the possibility
to define a default behaviour for this problem, as the default
configuration override all existing files (expect in sardine
implementation).

So next steps are/could be :
- The method putFile for the cmisService must be modified to work
 with this enhancement.
- Give the possibility to a user to change his default behaviour by
  editing portlet preferences.
- wait or implements the UI evolution, referenced here :
  https://github.com/Valums-File-Uploader/file-uploader/issues/40