Don't get "IMAGINARY" files type
Imaginary files are files that does not exist, so it's useless to show
them, see http://commons.apache.org/proper/commons-vfs/apidocs/org/apache/commons/vfs2/FileType.html#IMAGINARY
for more details on this file type.

In my case it correspond to broken symbolic links.

An other way to do would be to set these files as hidden, but we should
find an icon to associate.